backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
cors==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
bcrypt==4.1.2
PyJWT==2.10.1
python-multipart==0.0.6
